pyproject.toml
[tool.poetry]
name = "kombain"
version = "0.1.0"
description = ""
authors = ["hugyy85 <antonbat125a@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.5"
"pdfminer.six" = "^20200402"
tabula-py = "^2.1.0"


[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
